Added StandAlone.java.
This standalone file serves to be compiled by itself and generate Ri, Rm, U,
Xn, Xw. There will be NO Nx generated from this.

This will be a away to investigate the values of other variables without
messing with Nx

Last but most important feature: dt can be specified here. This is because
to compare values of u at 1-10^6, dt values must be adjusted appropriately.
Using the predicted dt from totaltime/100 is not longer effective.