Change how u calculated
Investigated error reveal that L cause significant error compared to gold
paper. This then prompt the change so that when u is NOT Inf, use the
regular equation, but when it does, use L

Also, added other things to print results into excel.

Next step from here is to investigate error of L putting on value of Xn and
Xw.

Cheers.