Complete implementing URange with U_Desire
During testing, we are interested in knowing values at a particular U.
Problem is my program use U as the dependent variable. So the only way to
guess U is through iterations.

Added a while loop to iterate. U_desire is the U I would like to get. Note
that the precision of U output depends on the time step dt. So I needed to
manually change dt to smaller values to reach a good precision output for U.